[native] Migrate Spark native tests to be inline with Presto's
